Alvgonfri

+++ src/main/java/org/springframework/samples/petclinic/game/GameRepository.java
@@ -7,10 +7,8 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import org.springframework.samples.petclinic.enums.State;
-
 @Repository
-public interface GameRepository extends CrudRepository<Game, Long>{
+public interface GameRepository extends CrudRepository<Game, Integer>{
 
     List<Game> findAll();
